Create main.py
A Python extension module called NumPy It offers quick and effective actions on collections of 
related data. Scipy is a Python library that is open-source and used for technical and scientific computing. OpenCV is a collection of programming tools with a real-time computer vision major focus. PyautoGUI is a Python-based, cross-platform GUI automation module. This allows you to 
automate computer chores by controlling the mouse and keyboard as well as performing simple picture 
recognition. 

 
Using Python, an image is created. First, it opens the camera and starts taking video. It then selects a frame and 
changes it to a grayscale image because it converts images to binary form, making it easy to find things. The 
face  is detected using Haar-cascade.  
 
Haar-cascade detects items from other photos after training on many positive and negative images. It will crop 
the frame and pass it on for additional processing after detecting the face. The Haar-cascade algorithm will 
detect eyeballs and trim the frame. Eye-cascade The Haar cascade recognises eyes.  
 
A numpy-supported four-variable array gives us x, y, w, and h all at once. The camera detects eyeballs from x 
and y; w is the width, and h is the height. These variables construct a rectangle around the eye and crop the 
image The rectangle begins at x and finishes at x+w horizontally and y and y+h vertically.